Create protoype with Bevy and a gRPC API (#3)
The Rust game engine Bevy has been set up with a gRPC API as a first proof of
concept. Bevy runs a main loop at 60Hz, and renders a window that displays the
current tick and the number of processes API requests. The API has been
implemented as a Bevy plugin, and runs a gRPC server in a background thread.
Incoming requests are pushed into a queue that is consumed during each tick in
the game engine. The response to each request is pushed back to the gRPC server
via a one-off async channel, unblocking the server to queue more requests while
waiting for a response.

This proof-of-concept is an experiment to get some hands on experience with
Bevy, but most likely not the final implementation for the integration between
Bevy and the API. Another idea that might be explored in the future is to use
a shared cache for entities, so that API requests don't have to wait for the
next tick in the game engine.