Switch to Godot Engine
The original plan was to use the new game engine Bevy to create Lunaria.
Bevy is written in Rust, which provides a good foundation for the
low-level services that Lunaria's API needs to run. But Bevy is still
very young and in active development, which makes it hard to find good
resources.

Godot on the other hand is more mature, and its community has created an
abundance of tutorials for all kinds of games. The engine is written in
C++, which means integrating the backend will be more difficult. But
behavior can be implemented using Rust, which should give us the best of
both worlds.

We have decided to experiment with Godot, since Bevy overwhelms us as
complete game development novices. As we gain experience and a better
understanding of the space, we might reconsider this decision. But for
now, we want to prototype as quickly as possible to learn if Lunaria can
even work as a game.