Set up grid and basic camera controls (#68)
A new scene has been created with a very simple grid. The main menu has
been reorganized and a button has been added to transition from the main
menu to this new scene. The scene renders a 64x64 grid of square fields,
and features a camera that can be moved around through key inputs.