Switch back to Bevy engine
After experimenting with Godot for a while, I have a theory that Bevy is
actually better suited for the tight integration that I want to achieve
between the engine and the game's API. This is an experimental branch to
quickly see how the core architecture would work with Bevy.

This commit replaces Godot with Bevy, and reimplements the main menu
that had been built with Godot.