Create event bus in Bevy engine (#82)
* Switch back to Bevy engine

After experimenting with Godot for a while, I have a theory that Bevy is
actually better suited for the tight integration that I want to achieve
between the engine and the game's API. This is an experimental branch to
quickly see how the core architecture would work with Bevy.

This commit replaces Godot with Bevy, and reimplements the main menu
that had been built with Godot.

* Create plugin for API

The API implementation has been moved into a custom Bevy plugin to
improve the code organization.

* Process commands in systems

A prototype has been implemented that connects the API and Bevy's
systems with each other. The API can receive commands and push them into
a message queue, from where systems can receive them and respond.